+++ packages/client/src/views/project/simple/images/Output.tsx
@@ -19,7 +19,10 @@ export const Output: React.FC<OutputProps> = ({ output }) => {
           <ImageActions file={output} type={ProjectFileType.OUTPUT} />
         </div>
       </div>
-      <ImagePreview className={styles.image} url={resolveImage(output.mrn)} />
+      <ImagePreview
+        className={styles.image}
+        url={resolveImage(output.mrn, undefined, true)}
+      />
     </div>
   );
 };